backend/pyproject.toml
[project]
name = "wabi-agent-backend-new"
version = "2.1.0"
description = "Gemini-powered LangGraph Orchestrator"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai",
    "openai>=1.59.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.30.0",
    "pillow>=10.0.0",
    "protobuf",
    "typing-extensions>=4.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.20.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "websockets>=11.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "langgraph-cli>=0.1.0",
    "langgraph-api>=0.7.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph_app"]
